pyproject.toml
[project]
name = "docstribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "motor>=3.6.0",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
